+++ src/lib/mealy.ml
@@ -192,7 +192,9 @@ module Make (S: Lts.STATE) = struct
 
   let map_state f s = { s with lts = M.map_state f s.lts }
   let map_attr f s = { s with lts = M.map_attr f s.lts }
+  let map_state_attr f s = { s with lts = M.map_state_attr f s.lts }
   let map_label f s = { s with lts = M.map_label f s.lts }
+  let map_transition f s = { s with lts = M.map_transition f s.lts }
 
   let clean s = { s with lts = M.clean s.lts }