+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## 0.6.1 (Feb XX, 2021)
     * fixed formatting bugs in generated doc 
+    * Added functions `Ltsa.map_state_attr`, `Ltsa.map_transition`, `Lts.map_state`
+      and `Lts.map_transition` and updated `examples/lts/map1` accordingly
 
 ## 0.6.0 (Sep, 16, 2019)
     * minor rewritings for ocaml >= 4.08